requirements.txt
django==2.2.12
beautifulsoup4==4.8.2
lxml
pillow==6.2.2
pyppeteer==0.0.25
loguru==0.4.1
djangorestframework==3.8.2